+++ lib/content-services/src/lib/common/mocks/ecm-user.service.mock.ts
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-import { PersonEntry, Person, PersonPaging } from '@alfresco/js-api';
+import { Person } from '@alfresco/js-api';
 
 export const fakeEcmUser: Person = {
     id: 'fake-id',
@@ -36,115 +36,4 @@ export const fakeEcmUser: Person = {
     userStatus: 'active',
     enabled: true,
     emailNotificationsEnabled: true
-};
-
-export const fakeEcmAdminUser = {
-    ...fakeEcmUser,
-    capabilities: {
-        isAdmin: true
-    }
-};
-
-export const fakeEcmUser2 = {
-    id: 'another-fake-id',
-    firstName: 'another-fake-first-name',
-    lastName: 'another',
-    displayName: 'admin.adf User',
-    email: 'admin.adf@alfresco.com',
-    company: null,
-    enabled: true,
-    emailNotificationsEnabled: true
-};
-
-export const fakeEcmUserNoImage = {
-    id: 'fake-id',
-    firstName: 'fake-first-name',
-    lastName: 'fake-last-name',
-    description: 'i am a fake user for test',
-    avatarId: null,
-    email: 'fakeEcm@ecmUser.com',
-    skypeId: 'fake-skype-id',
-    googleId: 'fake-googleId-id',
-    instantMessageId: 'fake-instantMessageId-id',
-    company: null,
-    jobTitle: null,
-    location: 'fake location',
-    mobile: '000000000',
-    telephone: '11111111',
-    statusUpdatedAt: 'fake-date',
-    userStatus: 'active',
-    enabled: true,
-    emailNotificationsEnabled: true
-};
-
-export const fakeEcmEditedUser = {
-    id: 'fake-id',
-    firstName: null,
-    lastName: 'fake-last-name',
-    description: 'i am a fake user for test',
-    avatarId: 'fake-avatar-id',
-    email: 'fakeEcm@ecmUser.com',
-    skypeId: 'fake-skype-id',
-    googleId: 'fake-googleId-id',
-    instantMessageId: 'fake-instantMessageId-id',
-    company: null,
-    jobTitle: 'test job',
-    location: 'fake location',
-    mobile: '000000000',
-    telephone: '11111111',
-    statusUpdatedAt: 'fake-date',
-    userStatus: 'active',
-    enabled: true,
-    emailNotificationsEnabled: true
-};
-
-export const fakeEcmUserList = new PersonPaging({
-    list: {
-        pagination: {
-            count: 2,
-            hasMoreItems: false,
-            totalItems: 2,
-            skipCount: 0,
-            maxItems: 100
-        },
-        entries: [
-            {
-                entry: fakeEcmUser
-            },
-            {
-                entry: fakeEcmUser2
-            }
-        ]
-    }
-});
-
-export const createNewPersonMock = {
-    id: 'fake-id',
-    firstName: 'fake-ecm-first-name',
-    lastName: 'fake-ecm-last-name',
-    description: 'i am a fake user for test',
-    password: 'fake-avatar-id',
-    email: 'fakeEcm@ecmUser.com'
-};
-
-export const getFakeUserWithContentAdminCapability = (): PersonEntry => {
-    const fakeEcmUserWithAdminCapabilities = {
-        ...fakeEcmUser,
-        capabilities: {
-            isAdmin: true
-        }
-    };
-    const mockPerson = new Person(fakeEcmUserWithAdminCapabilities);
-    return { entry: mockPerson };
-};
-
-export const getFakeUserWithContentUserCapability = (): PersonEntry => {
-    const fakeEcmUserWithAdminCapabilities = {
-        ...fakeEcmUser,
-        capabilities: {
-            isAdmin: false
-        }
-    };
-    const mockPerson = new Person(fakeEcmUserWithAdminCapabilities);
-    return { entry: mockPerson };
 };
